requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.32.3
youtube-transcript-api>=0.6.2
boto3>=1.34.0
